home: adds a home folder and a synchronisation command
From now on, trainer keeps activities synchronised in the home
folder. Activities can be synchronised from any directory and
consolidated in the home folder.

Synchronisation is pretty naive at the moment. This is a first
iteration which makes using the CLI much nicer.

This also opens the road to start implementing other activity
synchronisation mechanisms, and perhaps eventually gather activites
from third parties, or even automatically from the watch connected
to the computer.